♻️ refactor(test): testUtil 파일을 util 폴더로 이동
- 테스트 유틸리티 함수(mockQuestions, mockRandoms, getLogSpy)를 src/util/testUtil.js로 옮겼습니다.
- test 코드로서 실행되지 않게 하기 위해 __test__ 폴더 외부로 분리했습니다.
