+++ __tests__/ApplicationTest.js
@@ -1,5 +1,5 @@
 import App from '../src/App.js';
-import { getLogSpy, mockQuestions, mockRandoms } from './testUtil.js';
+import { getLogSpy, mockQuestions, mockRandoms } from '../src/util/testUtil.js';
 
 const runException = async (input) => {
   // given
