+++ __tests__/UserTest.js
@@ -1,6 +1,6 @@
 import { ERROR_MESSAGE } from '../src/constant/error.js';
 import User from '../src/User.js';
-import { getLogSpy, mockQuestions } from './testUtil';
+import { getLogSpy, mockQuestions } from '../src/util/testUtil.js';
 
 describe('User 클래스의 readPurchaseAmount 메서드 테스트', () => {
   beforeEach(() => {
